transaction-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-jose[cryptography]==3.3.0
aiokafka==0.11.0

